+++ Unreal/CarlaUE4/Plugins/Carla/Source/Carla/Vehicle/VehiclePhysicsControl.h
@@ -1,5 +1,6 @@
 // Copyright (c) 2017 Computer Vision Center (CVC) at the Universitat Autonoma
 // de Barcelona (UAB).
+// Copyright (c) 2019 Intel Corporation
 //
 // This work is licensed under the terms of the MIT license.
 // For a copy, see <https://opensource.org/licenses/MIT>.
@@ -8,6 +9,21 @@
 
 #include "Vehicle/WheelPhysicsControl.h"
 #include "VehiclePhysicsControl.generated.h"
+
+USTRUCT(BlueprintType)
+struct FGearPhysicsControl
+{
+  GENERATED_USTRUCT_BODY()
+
+  UPROPERTY(Category = "Gear Physics Control", EditAnywhere, BlueprintReadWrite)
+  float Ratio = 1.0f;
+
+  UPROPERTY(Category = "Gear Physics Control", EditAnywhere, BlueprintReadWrite)
+  float DownRatio = 0.5f;
+
+  UPROPERTY(Category = "Gear Physics Control", EditAnywhere, BlueprintReadWrite)
+  float UpRatio = 0.65f;
+};
 
 USTRUCT(BlueprintType)
 struct CARLA_API FVehiclePhysicsControl
@@ -43,6 +59,12 @@ struct CARLA_API FVehiclePhysicsControl
 
   UPROPERTY(Category = "Vehicle Engine Physics Control", EditAnywhere, BlueprintReadWrite)
   float ClutchStrength = 0.0f;
+
+  UPROPERTY(Category = "Vehicle Engine Physics Control", EditAnywhere, BlueprintReadWrite)
+  float FinalRatio = 1.0f;
+
+  UPROPERTY(Category = "Vehicle Engine Physics Control", EditAnywhere, BlueprintReadWrite)
+  TArray<FGearPhysicsControl> ForwardGears;
 
   // Vehicle Setup
   UPROPERTY(Category = "Vehicle Engine Physics Control", EditAnywhere, BlueprintReadWrite)